explicit typing for pauli string encodings
